Extract Python keymaster library from service
