+++ python/keymaster/README.md
@@ -1,0 +1,6 @@
+# Python Keymaster Library
+
+Reusable Python implementation of the Archon Keymaster business logic.
+
+This package is intended to hold the core wallet, DID, crypto, and asset
+behaviors independent of any specific HTTP service runtime.
