+++ python/keymaster/src/keymaster/__init__.py
@@ -1,0 +1,3 @@
+from .core import Keymaster, KeymasterError, UnknownIDError
+
+__all__ = ["Keymaster", "KeymasterError", "UnknownIDError"]
